refactor: extract Symphony orchestrator into packages/symphony
Splits the generic, zero-dependency coding-agent dispatch engine
(scripts/symphony.mjs, claude-symphony-channel.mjs, dispatch-channel.mjs)
out of js2wasm-specific scripts into a self-contained @loopdive/symphony
package under packages/symphony, following the existing packages/js2wasm
precedent.

- Split the monolith into lib/{yaml,workflow,dispatch-state,orchestrator,util}.mjs
  plus thin bin/{symphony,symphony-channel,symphony-dispatch}.mjs CLI entries.
- Fixed ROOT resolution to use process.cwd() instead of the script's own
  install location, required for it to work once out of scripts/.
- Fixed symphony-dispatch (formerly dispatch-channel.mjs) hardcoding
  plan/issues instead of reading tracker.issues_dir from WORKFLOW.md.
- Consolidated claims.json/messages.jsonl I/O, previously duplicated with
  slightly different shapes across all three scripts, into lib/dispatch-state.mjs.
- Added tests/symphony.test.ts unit-covering the extracted parsing/tracker/
  dispatch-state logic (none existed before).
- Rewired package.json scripts and .mcp.json to the new paths; trimmed
  plan/method/symphony-service.md to the js2wasm-specific wiring, with the
  generic architecture docs now living in packages/symphony/README.md.

✓
