+++ package.json
@@ -137,15 +137,15 @@
     "reconcile:tasks": "node scripts/reconcile-tasklist.mjs",
     "issues:pr-status": "node scripts/poll-merged-pr-issues.mjs --once --sync-artifacts",
     "issues:pr-status:watch": "node scripts/poll-merged-pr-issues.mjs --sync-artifacts",
-    "symphony": "node scripts/symphony.mjs",
-    "symphony:once": "node scripts/symphony.mjs --once",
-    "symphony:dry-run": "node scripts/symphony.mjs --once --dry-run",
-    "symphony:status": "node scripts/symphony.mjs --status",
-    "symphony:control": "node scripts/symphony.mjs --control",
-    "dispatch:queue": "node scripts/dispatch-channel.mjs queue",
-    "dispatch:request-claude": "node scripts/dispatch-channel.mjs request-claude",
-    "dispatch:inbox": "node scripts/dispatch-channel.mjs inbox",
-    "claude:symphony-channel": "node scripts/claude-symphony-channel.mjs",
+    "symphony": "node packages/symphony/bin/symphony.mjs",
+    "symphony:once": "node packages/symphony/bin/symphony.mjs --once",
+    "symphony:dry-run": "node packages/symphony/bin/symphony.mjs --once --dry-run",
+    "symphony:status": "node packages/symphony/bin/symphony.mjs --status",
+    "symphony:control": "node packages/symphony/bin/symphony.mjs --control",
+    "dispatch:queue": "node packages/symphony/bin/symphony-dispatch.mjs queue",
+    "dispatch:request-claude": "node packages/symphony/bin/symphony-dispatch.mjs request-claude",
+    "dispatch:inbox": "node packages/symphony/bin/symphony-dispatch.mjs inbox",
+    "claude:symphony-channel": "node packages/symphony/bin/symphony-channel.mjs",
     "check:issue-ids": "node scripts/check-issue-ids.mjs",
     "check:issue-ids:staged": "node scripts/check-issue-ids.mjs --staged",
     "check:issue-ids:against-main": "node scripts/check-issue-ids.mjs --against-main",
